restructured files for use with dep
